mejoras en navbar

+++ api/src/controllers/videogames/postVideogame.js
@@ -4,35 +4,54 @@ const fs = require('fs')
 
 const postVideogame = async (id, name, description, platforms, image, imageBase64, released, rating, genres) => {
     // Decodificar la cadena base64 de la imagen
-    
+
     const base64Data = imageBase64.replace(/^data:image\/\w+;base64,/, '');
     const dataBuffer = Buffer.from(base64Data, 'base64');
 
-    const fileName = `${Date.now()}.jpg`;
+    const fileName = image //`${Date.now()}.jpg`;
 
     // Ruta donde se guardará el archivo
     const images = '../../../../images/'
     const filePath = path.join(__dirname, images, fileName);
 
-
+    let edited = false
     try {
-
-        fs.writeFile(filePath, dataBuffer, (error) => {
-            if (error) {
-                return ({ error: error.message })
-            }
-        })
-
-        const [videogame, created] = await Videogame.findOrCreate({
+        if (!fs.existsSync(filePath)){
+            fs.writeFile(filePath, dataBuffer, (error) => {
+                if (error) {
+                    return ({ error: error.message })
+                }
+            })
+        }
+        const newData = {
+            name: name,
+            description: description,
+            image: fileName,
+            released: released,
+            rating: rating
+        }
+        let [videogame, created] = await Videogame.findOrCreate({
             where: { id: id },
-            defaults: {
-                name,
-                description,
-                image: fileName,
-                released,
-                rating
-            }
+            defaults: newData
         })
+        //lo encontro
+        if (!created) {
+            edited = true
+            await Videogame.update(newData, {
+                where: { id: id }
+            })
+            //borro generos y platforms 
+            await Videogame_genre.destroy({
+                where: {
+                    videogameId: id
+                }
+            })
+            await Videogame_platform.destroy({
+                where: {
+                    videogameId: id
+                }
+            })
+        }
 
         //mapeo y envio Videogame_genre
         const videogameGenres = genres.map(async (g) => {
@@ -56,9 +75,8 @@ const postVideogame = async (id, name, description, platforms, image, imageBase6
 
         const platformResults = await Promise.all(videogamePlatforms);
 
-        return { videogame, created, genres: genreResults, platforms: platformResults }
+        return { videogame, created, edited, genres: genreResults, platforms: platformResults }
     } catch (error) {
-        console.log(error)
         return ({ error: error.message })
     }
 }
